docs(ngsctp): add interop testing plan and partial type fixes
- Added `docs/NG_SCTP_TODO.md` detailing the blocking compilation
  issues in the `ng-sctp` module and providing a concrete plan for
  testing interoperability with the Pion SCTP reference implementation.
- Applied partial fixes to `NgChunk.kt`, `CongestionModel.kt`, and `IoUringSctpTransport.kt`
  to resolve unsigned type and `java.nio` mismatches.
- Note: `ng-sctp` remains excluded from `settings.gradle.kts` until
  transport implementations are completed and all compilation errors
  are resolved.

+++ libs/ng-sctp/src/commonMain/kotlin/dev/jnorthrup/ngsctp/ml/CongestionModel.kt
@@ -232,7 +232,7 @@ class CubicCongestionPredictor : CongestionPredictor {
         // On loss, reduce to (1-beta) * cwnd
         wMax = features.cwnd
         lastLossTime = System.currentTimeMillis()
-        return (features.cwnd * (1 - BETA).toUInt()).coerceAtLeast(4380u)
+        return (features.cwnd.toDouble() * (1 - BETA)).toUInt()).coerceAtLeast(4380u)
     }
 }
